Rename GitHub wizard file names
