+++ apps/console/src/features/identity-providers/components/wizards/github/github-authentication-provider-create-wizard-content.tsx
